backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=41.0.0
bleach>=6.1.0
celery==5.3.4
redis==5.0.1
httpx>=0.24.0,<0.25.0
requests>=2.31.0
email-validator>=2.1.0.post1
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1
python-dotenv==1.0.0
websockets==12.0
supabase==2.0.0
postgrest==0.13.0
